I added a default node for returning in a parse call. Although I added the majority of the flexibility in the last commit, a few were done in this as well.
Although this would be obvious if the reader read how the loop node works, if you use return inside of a loop, it does not end parsing in the entire template, but merely the loop iteration. If it is preparsed, it removes the excess content and then goes through the iterations, but if not, it is the same process repeated multiple times (Not that there should be any reason why return is not preparsed). As this was one of the last things I had planned for the immediate future of SUIT, I plan to create a new release in the following days.