Helper has been merged into the main file. SUIT is becoming smaller and smaller, and helper module seems superfluous. Although in PySUIT, it isn't proper to define so many functions as methods, I plan to remove the wrapper soon anyway, so it won't matter. Also, I have removed the default config for nodes to it because the average user doesn't need to modify the predefined nodes, and if they do, it is easier than defining all of them. Last, I added a blacklist / whitelist to [var].
Next, I plan to remove the class wrapper from PySUIT, have gettemplate removed and replace it with a node that retrieves a template and a node that retrieves code (Code no longer seems to translate templates, but to define variables), and possibly remove the paths and filetypes for [template] for a whitelist / blacklist.

In the end, I hope to have it so that SUIT can be used without any configuration so class initialization, at least, for PySUIT, as the class wrapper is necessary due to the lack of namespaces. In order to remove the wrapper, I may use the suit module for suit.vars, and for consistency, in PHP SUIT, the class instead of the suit->vars array.