Well, that was quick. The class wrapper for PySUIT has been removed, and in PHP SUIT, the class is used for variables. gettemplate has been removed, and there is now a [code] node which includes / imports a code file, and the [template] tag simply gets the contents of the provided file. Lastly, you can now whitelist / blacklist these files. I have included loops that whitelist the files for all of the example programs and TIE.
Next, I need to do some cleanup with the structure of parse (parseconfig, parsecache, and parsepositions? Really?), work on the example template, and really try to get some docs done (I will be rewriting all of my sites to PySUIT for various reasons. Hopefully, I can release 1.3.5 soon, and can work on my new project, SLACKS, which stands for SLACKS Lets Application Coders Know SUIT. I will write it for both PHP and PySUIT, and it will hopefully replace the only important part of TIE, allowing me to discontinue it, and if anyone's ever read it, that's a good thing. :)