SLACKS (SLACKS Lets Application Coders Know SUIT) is a far more advanced version of TIE's debugger. I plan to have it so it diagrams the entire transformation, recursively. It will be Javascript intensive, and will be able to be setup onto any script relatively. You will be able to debug SLACKS using SLACKS because the profiling is sent through JSON. So far, I have only gotten the profiling done and the form that submits it.
Next, I need to clean up parse and start on the display aspect. I hope for it to have tabbing so that you can open and close each box (It will automatically close when said parse is done to avoid a gigantic block of boxes), indentation for each recurse, manual and slide show capabilities, scroll control (As in, should you scroll the bar to view other boxes, or should it auto scroll? Especially helpful in the slide show mode, but it certainly should be an option), and highlighting for the ignored and taken ranges, which are now automatically logged, regardless if you are preparsing or not.