I did not previously mention this, but I added a malformed mode for parsing where non-matching tags would parse anyway. In this revision, I have extended it to parse unopened / unclosed strings. I don't support weak parsing, but in some cases, such as BBCode, a person might want to use it. More practically, if you know you are going to be given a malformed document, using the new extension, you can repair it. Ex. If you are parsing HTML, and you have it so that every match is replaced with the opening string + the case + the closing string, and you are given the following string:
    Text
    </div>
</div>

It would parse to:

<div><div>
    Text
    </div>
</div>

Which could prove useful in some situations. Regardless, it is off by default.

The main reason I added this was because I thought it would be necessary in order to have an HTML Mode in SLACKS and not completely break the page. That is still debatable. Speaking of which, I was experimenting with having the Javascript generate the iterations, but my methods involved flattening the array, and were overall very unconventional. So, I decided that I would have it so that SUIT generates the whole thing, and if Javascript is enabled, all the necessary boxes will be hidden until you use the Next and Previous buttons to go through the process (There will also be some sort of slideshow format).

Today, I tried to do the actual iteration generation. In order to do this, I would need to create some way to do recursive functions in SUIT. The current way to do "functions" in SUIT is to store a template and parse it in multiple situations with different variables. So, I figured I could make it recursive by nesting a template in itself until there are no longer any recursions.

Somewhere along the line, I realized that node preparsing had to be removed. The entire concept of a recursive function is that you are running the same function over and over again with different variables. As node preparsing would loopvars to stay constant in order for consistency, that meant that the variables would never change, and the loop would go on infinitely. So, I did away with it. Preparsing should only be extended when it does not destroy consistency and helps save time in many situations, and this obviously was not the case for node preparsing.

At first, I noticed that PHP was exhausting itself, so I figured there was something wrong with nesting more than 2 loops. So, I debugged, and finally decided to test it on Python, where I figured it would be more efficient. It generated after a while. That is when I learned that Python seemingly does not have any memory limits by default. I was contacted by my host that I used 320MB!? So, I'm assuming that something went horribly wrong. I believe that at this point, SUIT, and by extension its looping / preparsing system, is relatively efficient. Perhaps, however, not efficient enough to handle recursive functions. Hopefully, I can make it so these operations are acceptable.