The move to branches is going relatively smoothly. The remaining nodes to convert are return and trim. I am pretty sure there will have to be some way for the inner nodes to communicate with the outer nodes, probably by returning information. A developer friend of mine said that preparsing is pretty hacky and unnecessary for a little amount of optimization, and I was having second thoughts as well, so it has been removed. He also mentioned that my terminology was wrong, so I put the position generation in parse to tokens, the tree generation to parse, and parse to execute. Loop and condition no longer use skip. This is through a new configuration called 'treefunctions' which executes before walking through the items. The functions that execute on the actual string items have been renamed to 'stringfunctions'. That should be enough, unless I decide to have the returning thing return functions, which is what I am leaning towards because I'm not sure how the outer nodes will use the data otherwise.
My test to see if these optimizations have done the job will be having SLACKS generate fieldsets for every branch of the tree using a recursive templating function. If that works, I will convert to Python, port to the other packages, and proceed to log the items executing. That should be more efficient than my last model, and would require very little extra work.

Lastly, I forgot to mention in the last commit that parsing unopened strings in malformed mode is ambiguous, so it has been removed. There are also two different keys to handle the other malformed cases, 'mismatched' and 'unclosed', so that users can enable one without the other.