I have converted return and trim, so now the PHP SUIT conversion to trees has been completed. One of the things changed is that previously, each individual pieces of the tree were transformed by the string functions, and now I have it so it transforms the combined results, as the earlier version wouldn't work in several cases, and is probably more intensive.
There are now three groups of functions:

treefunctions - Functions that run before walking through the tree that tend to modify the tree.
returnfunctions - Functions returned from the inner nodes that tend to modify the tree. This is used as an alternative of modifying the stack now that the stack is multidimensional and harder to manipulate.
stringfunctions - Functions run after the tree has been walked through the tree and tend to modify the imploded result.

In addition to return functions, there is also the return var. When in a lower function, if you assign $params['returnvar'] to something, it will be accessible by the outer nodes as $params['returnedvar'] in the return functions. It is essentially return function's version of 'var', although 'var' is accessible in all cases whereas 'returnedvar' is only accessible in return functions. 'returnvar' can be modified in all three function groups.