I have changed a bit in the naming of variables and files. $return wasn't specific enough, so in most cases I replaced it with strings. Also, pyslacks is now slacks because if you are using Python, it'd be redundant to have py as a prefix (Unless it's clever like Pylons). In addition to this, I attempted at removing monotony from my files. Both suit.py and nodes.py validate with pylint.
In the SUIT and PySUIT examples, I have put a box that shows the contents of variables.inc.php and variables.py respectively. For both cases, I have made it so the code is highlighted when possible.

Lastly, I have realized that [escape] doesn't really escape, but just skips. So, I converted it to [skip], and added an [escape] node that escapes a list of strings using suit.escape. Not a huge amount of functionality, but I figure suit.escape should be used in the implementation at least once. Hopefully I can go onto documenting the important functions soon enough.