PySUITlons has had various changes.
(1) The default SUIT nodes are no longer re-created from scratch, as the SUIT nodes structure changes quite often and is easy to miss a few things.

(2) For code files, you are responsible for importing all relevant modules that are going to be used yourself. "Explicit is better than implicit"

(3) There are less changes required now to the project structure in comparison to before, as I feel that __init__ files that contain code are rather messy.

The current issue right now is the JSON encoding of SQLAlchemy results, although that shall be sorted in the next commit.