I have added compatibility for nodes that have the same opening string as its closing string, or flat nodes. I call them flat nodes because you can not nest them for obvious reasons. When it hits the string, it will simply open if it is not opened, and close if it is.
Next, I have removed jsonencode in PySUIT. It is hard to know what the user wants to do in these situations, so I will leave it to them to convert. Objects will still be compatible in every other context.

Lastly, I made it so the HTML Mode boxes in SLACKS are commented out until I have the solution.