I want to say in advance that this is a huge commit, so if I forget something, I will hopefully mention it in a later one.
Firstly, I converted the [code] tag in BBNode to use <pre>. This will handle tabs, which I feel is important.

Secondly, in order to display SLACKS in a logical way, I moved towards making it so that the contents of a tree do not transform, but that they compile a case which they transformed after all iterations have been walked through. This method seems more consistent with the original tree-less model.

In order to accommodate this, I decided that the user should be editing the case on the branch instead of a separate arbitrary variable. So, params['tree'] can be edited by all functions called by walk. params['case'] and params['node'] can be accessed through params['tree']['case'] and params['tree']['node'] respectively.

Because of this change, the function group names do not make much sense. treefunctions do not necessarily edit the tree, and stringfunctions do not necessarily edit strings. In most cases, treefunctions, stringfunctions, and returnfunctions either modify the tree or terminate walking through it. So, treefunctions have been renamed to prewalk, and stringfunctions have been renamed to postwalk, which I did because that is when they run. The name for returnfunctions has not been renamed because of its context and that one of the other function groups returns it to a particular iteration on the tree. It does not exactly happen during the walk, but instead happens in an outer layer, so I figured that I should just keep it that way, similarly with returnvar and returnedvar.

Lastly, I got SLACKS to display the data in a relatively static way. I have provided three buttons that shows the different steps of each walk: Before, Tree (Generation), and After. In the end product, it will have buttons to go step by step and change the phase for each individual box. The current model does not work because pressing After would collapse all of the inner boxes instead of analyzing its transformation, but I plan on changing it as soon as possible. In the end, I am not 100% sure that it will remain a web application due to the massive amount of output (10,000 lines on the example). Either way, the web version would be a suitable model, so I hope I can get it up and running ASAP.

1.3.5: When it's done. I sure hope it's reasonably soon, but soon I'll have nothing else but documentation, and I will actually attempt at doing it.