I have completed the model for SLACKS. In it, you can toggle between phases in the tree and branches. All boxes have the phases "Tree" and "Case", but the highest level ones also have an "Original" one. If execute is run at the same time as a certain branch is transforming, the branch will link to it in the "Parallel" box. The boxes are ordered by the outermost box's id.
It seems like a necessity to convert SLACKS to Flash for several reasons:

1. 10,000 lines of markup will often crash a browser.
2. SUIT can handle recursive functions fairly well, but in extreme cases like this, there is a time delay.
3. The UI is relatively overwhelming, and Flash could potentially show it in a much better way, possibly isolating the specific trees. In addition, it could probably format the text as HTML without breaking the document, which is something I've always wanted to allow.
4. It would allow me to make the SLACKS engine in one language as opposed to two.

Expanding on 3, I think I will just provide SLACKS as an application of suitframework.com and provide buttons on all of my applications to post to it. It wouldn't be particularly hard to setup buttons on any (Py)SUIT site either. Either version could post to it and get the necessary output, so there is no reason for me to develop it with two different wrappers. For educational purposes, I plan to continue committing SLACKS. As I believe, if done correctly, SLACKS will be a very popular feature (For both entertainment and debugging), it should hopefully give the site the traffic it needs.

I have no experience with Flash, so I have no idea when I will make this conversion, but I will attempt to learn it for this. Volunteers are welcome.