As I have been working on the new site, I have been making structural changes to the packages. SUIT will now be packaged as a module, and the default rules, BBNode, and the rule modifications for Pylons, will be in a package called Rulebox. In addition, to remove version bias, SUIT will now be referred to as SUIT for PHP (Packaged as suit-php), and PySUIT will be referred to as SUIT for Python (Packaged as suit-python).
Next, I have converted the default separator from => to . for consistency with the SUIT for Pylons project. It makes more sense, anyway.

Also, I have removed SLACKS and the examples for the repository. Both of these things will be integrated into the new suitframework.com, and I will more than likely open-source the website when I complete it.

In addition, all packages now print the JSON encoded tree when adding ?slack=true to the URL. The replacing method was inefficient and would not allow people to share the result via a URL.

Lastly, I have acknowledged that there isn't, and probably never will be, a SUIT Group. So, SUIT is now Copyright Brandon Evans and Chris Santiago. Chris has done a lot for SUIT, including the entire SUIT for Pylons project, and I haven't credited him as often as I should have because I am a jerk like that. The new documentation will still contain a credits page, but it may be trimmed. The only real contribution beyond us is Alex Munroe, and he didn't work with the code as much as answering my plethora of questions.

Still am not sure what I am going to do about the versioning. I would be willing the rename the old packages if necessary. I just want the different language versions to have the same version number, and to somehow emphasize that the new release is a huge step up from 1.3.4. I really wish I started with SVN from the beginning.