I have decided that I am going to use the major.minor.release format for versioning. I will not be modifying the old packages. So, SUIT and TIE will be at 2.0.0, while Rulebox will be at 1.0.0.
In addition, I have fixed an error which occurs when someone provides an unclosed opening string. Now, instead of triggering an error, it will just concatenate the opening string with the contents. This worked before, but I forgot to keep it consistent with the other upgrades.

Next, I have fixed several bugs with TIE. Yeah, I know, who cares?

Lastly, I would like to mention that in one of the latest commits, I did some fixes to the caching. It only caches the necessary elements now, so it saves memory usage. Also, caching the functions in PySUIT would often cause it to bug out because the address of the functions kept changing. Also, the cache uses JSON instead of serialization now.