-bbcode.class.php --Fixed the example. -templating.class.php --You can now have an attribute named "0". Before, it treated this as null. -suit.class.php --Just like above, just executing "0" should now work correctly.
I hope to eventually overhaul this code base and work on the documentation. I'm constantly surprised on how well SUIT is holding together considering how old it is, but that doesn't mean it's perfect, and instances like this prove that I need to re-examine it.

First commit in over 2 years! I ain't dead yet! :P